pyproject.toml
[tool.poetry]
name = "tuto-django"
version = "0.0.1"
description = "Tuto for Django"
authors = ["Sébastien CHAZALLET <pythoniste@protonmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
Django = "~5"
bpython = "^0.24"
psycopg2-binary = "^2.9.9"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.2"
pillow = "^10.3.0"
pyparsing = "^3.2.5"
pydot = "^4.0.1"
pydotplus = "^2.0.2"
pygraphviz = "^1.14"
pygments = "^2.19.2"
django-rosetta = "^0.10.2"
djangorestframework = "^3.16.1"
markdown = "^3.9"
django-filter = "^25.2"
django-ninja = "^1.4.5"
django-mptt = "^0.18.0"
django-polymorphic = "^4.1.0"
django-jazzmin = "^3.0.1"
django-bootstrap5 = "^25.2"
django-bootstrap-icons = "^0.9.0"
django-datatable-view = "^2.1.6"
martor = "^1.7.15"
django-crispy-forms = "^2.4"
crispy-bootstrap5 = "^2025.6"
celery = {extras = ["amqp"], version = "^5.5.3"}
pika = "^1.3.2"
amqp = "^5.3.1"
django-celery-results = "^2.6.0"
django-celery-beat = "^2.8.1"
factory-boy = "^3.3.3"
faker = "^37.12.0"
behave-django = "^1.7.0"
pyhamcrest = "^2.1.0"
beautifulsoup4 = "^4.14.2"
django-health-check = "^3.20.0"
psutil = "^7.1.3"
django-redis = "^6.0.0"
bcrypt = "^5.0.0"
gunicorn = "^23.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
